requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
gunicorn==23.0.0
whitenoise==6.9.0
uuid6==2025.0.1
dj-database-url==3.0.1
psycopg2-binary==2.9.11
requests==2.33.1